Value text object to sch for D1, D2, Q1, Q2, Q3
I used part #s for values (rather than the "values") for these
components. This is to ensure production (and users) will know when they
are a different component, when looking at the schematic. Note they are
text objects living in the wild, and not actually part of the device in
the library. Although, the values are inside the properties for each
device as well.