switched to "TIGHT" DRU, bumped vias up to 15 mil
Note, when I moved to the TIGHT DRU the restring for vias is set to 7
mil (this makes for a tighter annular ring). The previous DRU settings
had the restring set at 10 mil, and all of the vias were 12 mil drills.
When I changed to a 7 mil restring and a 15 mil drill, the vias were
still a smaller footprint than in the original. So ultimately, this
didn't necessitate any component moving or trace re-routing.